pyproject.toml
[project]
name = "mcp-obsidian-sse"
version = "0.1.0"
description = "MCP server to work with Obsidian via the remote REST plugin"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "mcp>=1.6.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.31.1",
]

[[project.authors]]
name = "Duncan Chiu"
email = "chiuduncan409@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
]